Change preview file detectino method
Now it is checking for a second png.
It will use the first one as the highres one and the second one as the preview. Order is in alphabetical order meaning icon.png is ahead icon_preview.png so if the files are named correctly the correct files will get picked.